[wip] kickass factoring of chart rendering in the charts crate

+++ libs/charts/src/chart/time/size.rs
@@ -5,7 +5,7 @@ use crate::common::*;
 use point::TimeSizePoints;
 
 /// Initial size value.
-const INIT_SIZE_VALUE: usize = 0;
+const INIT_SIZE_VALUE: u32 = 0;
 
 /// Total size over time chart.
 #[derive(Debug, Serialize, Deserialize)]
@@ -13,9 +13,9 @@ pub struct TimeSize {
     /// Timestamp of the last allocation registered.
     timestamp: SinceStart,
     /// Current total size.
-    size: PointVal<usize>,
+    size: PointVal<u32>,
     /// Map used to construct the points.
-    map: Map<Date, PointVal<(usize, usize)>>,
+    map: Map<Date, PointVal<(u32, u32)>>,
 }
 
 impl TimeSize {
@@ -56,7 +56,7 @@ impl TimeSize {
     }
 
     /// Initial size.
-    fn init_size_point(filters: &filter::Filters) -> PointVal<usize> {
+    fn init_size_point(filters: &filter::Filters) -> PointVal<u32> {
         PointVal::new(INIT_SIZE_VALUE, filters)
     }
 }
